refactor: replace styles/styled with @styles alias
