+++ src/context/theme.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createContext, useState } from 'react';
-import { StyledThemeProvider, SiteTheme } from 'styles/styled';
+import { StyledThemeProvider, SiteTheme } from '@styles/styled';
 import { theme } from 'styles/theme';
 import { themeBackgroundColor } from 'styles/globals';
