+++ src/screens/Blog/PageTemplate.tsx
@@ -17,7 +17,7 @@ import { MenuBar } from 'components/MenuBar';
 import { Frontmatter } from 'types';
 import SEO from 'components/SEO';
 import { ThemeProvider } from 'context/theme';
-import { styled, media, SiteTheme } from 'styles/styled';
+import { styled, media, SiteTheme } from '@styles/styled';
 import { SideMenu } from 'components/SideMenu';
 import { joinSubtitleAndDescription } from 'utils/seo';
 import { AvailableTranslations } from '@screens/Blog/components/AvailableTranslations';
