+++ src/screens/Blog/components/AvailableTranslations/styled.ts
@@ -1,4 +1,4 @@
-import { styled } from 'styles/styled';
+import { styled } from '@styles/styled';
 
 export const Phrase = styled.p`
   margin: 0;
