+++ src/screens/Blog/components/MdxComponents/Gif/Gif.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { styled } from 'styles/styled';
+import { styled } from '@styles/styled';
 
 type GifProps = {
   src: string;
