+++ src/screens/Home/components/Posts.tsx
@@ -3,7 +3,7 @@ import { FormattedMessage } from 'react-intl';
 import { motion, AnimatePresence } from 'framer-motion';
 
 import { InfiniteScroll } from 'components/InfiniteScroll';
-import { styled } from 'styles/styled';
+import { styled } from '@styles/styled';
 import { PostEdge, PostEdges } from 'types';
 import { PostCard } from 'components/PostCard';
 import { PostFilters } from '@screens/Home/types';
