+++ src/types/styled-components.d.ts
@@ -1,6 +1,6 @@
 // import original module declarations
 import 'styled-components';
-import { SiteTheme } from 'styles/styled';
+import { SiteTheme } from '@styles/styled';
 import { CvTheme } from 'components/CV/styles';
 
 // and extend them!
